Fix ts field to be epoch milliseconds.
This is fixed in both the eBPF module and in the user-space code.

eBPF change:

Previously the eBPF code used bpf_ktime_get_ns() which returns
a CLOCK_MONOTONIC value which is the number of nanoseconds since the
kernel was booted (but doesn't include suspended time). This is changed
to use bpf_ktime_get_boot_ns() which return CLOCK_BOOTIME. This is
similar but also includes time the that the kernel was suspended (so
actual nanoseconds since system booted). We would like to use something
like bpf_ktime_get_tai_ns() to get CLOCK_TAI values (like
CLOCK_REALTIME but without discontinuities) but that function didn't
appear until kernel 6.1
(https://github.com/iovisor/bcc/blob/master/docs/kernel-versions.md).

User-space change:

We calculate the millisecond epoch time of the kernel boot when the
listener starts up. Then we use add that to the ts value in the event
(which is the number of epoch nanosecond since boot). This gives us an
epoch time in milliseconds which is what we print out in the ts field.